refactor(separation): move lifecycle orchestration to services
Extract services::separation for start_job/batch, status DTOs, and
bootstrap/publish seams. Thin commands/separation and batch_separation
to IPC adapters (C5).
