+++ src-tauri/src/services/mod.rs
@@ -1,3 +1,4 @@
 pub mod cdg;
 pub mod playback;
 pub(crate) mod playback_source;
+pub mod separation;
